BAEL-1510 - RxJava - Combining Observables
